+++ noodles-sam/src/reader/record.rs
@@ -10,8 +10,8 @@ mod sequence;
 mod template_length;
 
 pub(crate) use self::{
-    cigar::parse_cigar, flags::parse_flags, mapping_quality::parse_mapping_quality,
-    sequence::parse_sequence, template_length::parse_template_length,
+    cigar::parse_cigar, flags::parse_flags, sequence::parse_sequence,
+    template_length::parse_template_length,
 };
 
 use std::{
@@ -20,8 +20,9 @@ use std::{
 };
 
 use self::{
-    data::parse_data, position::parse_alignment_start, quality_scores::parse_quality_scores,
-    read_name::parse_read_name, reference_sequence_id::parse_reference_sequence_id,
+    data::parse_data, mapping_quality::parse_mapping_quality, position::parse_alignment_start,
+    quality_scores::parse_quality_scores, read_name::parse_read_name,
+    reference_sequence_id::parse_reference_sequence_id,
 };
 use super::read_line;
 use crate::{alignment::Record, Header};